xgt-kafka-producer/requirements.txt
psycopg2-binary
kafka-python
python-dotenv